implemented feature request: Monitor hosts in OK state to perform actions #128
see: https://github.com/HenriWahl/Nagstamon/issues/128
Added filter to show all hosts which are up and a filter to show all
services which are ok. 
If selected:
You could see all hosts or services which are "ok".
You can select actions on them.
The status bar will contain a green "OK" field, with the number of hosts
and services.
The web browser will open with all hosts or services, not only
the "wrong" ones.
The default is "filtered", so the old behaviour will consist.
Only tested for nagios, code for the others is changed, but maybe
the "up" hosts and "ok" services may not been shown.